save data when file is forcefully closed
Before, the program would not save any data if it crashed.
Also implemented a force_cache option. The cache is only flushed when a change in the number of repositories is detected, so if a user has added and deleted an equal number of repositories since the last time the cache was built, the LOC count will not be accurate. If enabled, force_cache will properly update the cache file.